# backend/src/main/kotlin/com/kangsiwoo/whenioff/route/api/CommuteRouteController.kt
package com.kangsiwoo.whenioff.route.api

import com.kangsiwoo.whenioff.route.application.CommuteRouteService
import com.kangsiwoo.whenioff.route.application.RouteCalibrationService
import com.kangsiwoo.whenioff.route.application.RouteLegService
import jakarta.validation.Valid
import org.springframework.http.HttpStatus
import org.springframework.web.bind.annotation.GetMapping
import org.springframework.web.bind.annotation.PatchMapping
import org.springframework.web.bind.annotation.PathVariable
import org.springframework.web.bind.annotation.PostMapping
import org.springframework.web.bind.annotation.PutMapping
import org.springframework.web.bind.annotation.RequestBody
import org.springframework.web.bind.annotation.RequestMapping
import org.springframework.web.bind.annotation.ResponseStatus
import org.springframework.web.bind.annotation.RestController

@RestController
@RequestMapping("/api/v1/commute-routes")
class CommuteRouteController(
    private val commuteRouteService: CommuteRouteService,
    private val routeLegService: RouteLegService,
    private val routeCalibrationService: RouteCalibrationService,
) {
    @GetMapping
    fun list(): List<CommuteRouteResponse> = commuteRouteService.list()

    @PostMapping
    @ResponseStatus(HttpStatus.CREATED)
    fun create(
        @Valid @RequestBody request: CreateCommuteRouteRequest,
    ): CommuteRouteResponse = commuteRouteService.create(request)

    @GetMapping("/{id}")
    fun detail(
        @PathVariable id: Long,
    ): CommuteRouteDetailResponse = commuteRouteService.detail(id)

    /** 이름·사용 여부·기본 목표 도착 시각·대상 day_type 일부 수정 (#68). 내 경로가 아니거나 없으면 404. */
    @PatchMapping("/{id}")
    fun update(
        @PathVariable id: Long,
        @Valid @RequestBody request: UpdateCommuteRouteRequest,
    ): CommuteRouteResponse = commuteRouteService.update(id, request)

    @PutMapping("/{id}/legs")
    fun replaceLegs(
        @PathVariable id: Long,
        @Valid @RequestBody request: ReplaceRouteLegsRequest,
    ): CommuteRouteDetailResponse = routeLegService.replaceLegs(id, request.legs)

    /** 경로의 보정값과 샘플 수 (#60). 내 경로가 아니거나 없으면 404. 읽기 전용. */
    @GetMapping("/{id}/calibration")
    fun calibration(
        @PathVariable id: Long,
    ): RouteCalibrationResponse = routeCalibrationService.calibration(id)
}