# backend/src/main/kotlin/com/kangsiwoo/whenioff/route/api/RouteDtos.kt
package com.kangsiwoo.whenioff.route.api

import com.kangsiwoo.whenioff.common.domain.DayType
import com.kangsiwoo.whenioff.route.domain.CommuteDirection
import com.kangsiwoo.whenioff.route.domain.CommuteRoute
import com.kangsiwoo.whenioff.route.domain.LegType
import com.kangsiwoo.whenioff.route.domain.RouteLeg
import com.kangsiwoo.whenioff.route.domain.RouteLegSignalCrossing
import com.kangsiwoo.whenioff.transit.api.TransitLineResponse
import com.kangsiwoo.whenioff.transit.api.TransitStopResponse
import jakarta.validation.Valid
import jakarta.validation.constraints.DecimalMax
import jakarta.validation.constraints.DecimalMin
import jakarta.validation.constraints.NotBlank
import jakarta.validation.constraints.NotEmpty
import jakarta.validation.constraints.Pattern
import jakarta.validation.constraints.Positive
import jakarta.validation.constraints.Size
import java.time.Instant
import java.time.LocalTime

data class CreateCommuteRouteRequest(
    @field:NotBlank val name: String,
    val direction: CommuteDirection,
    @field:DecimalMin("-90.0") @field:DecimalMax("90.0") val originLat: Double,
    @field:DecimalMin("-180.0") @field:DecimalMax("180.0") val originLng: Double,
    @field:DecimalMin("-90.0") @field:DecimalMax("90.0") val destinationLat: Double,
    @field:DecimalMin("-180.0") @field:DecimalMax("180.0") val destinationLng: Double,
    val isActive: Boolean = true,
    /** 일괄 추천(#68)의 기본 목표 도착 시각, KST 벽시계 (`09:00` / `09:00:00`). 없으면 일괄 추천에서 빠진다. */
    val defaultTargetArrivalTime: LocalTime? = null,
    /** 일괄 추천을 돌릴 day_type 집합. 생략하면 `[WEEKDAY]`, 빈 배열은 400. */
    @field:Size(min = 1) val defaultTargetDayTypes: Set<DayType>? = null,
)

/**
 * 경로 일부 수정 (#68). 필드가 null(생략)이면 그대로 둔다. 목표 시각은 null이 "그대로"라서
 * 지우려면 `clearDefaultTargetArrivalTime: true`를 보낸다 (시각과 같이 보내면 400).
 */
data class UpdateCommuteRouteRequest(
    @field:Pattern(regexp = "(?s).*\\S.*", message = "must not be blank") val name: String? = null,
    val isActive: Boolean? = null,
    val defaultTargetArrivalTime: LocalTime? = null,
    val clearDefaultTargetArrivalTime: Boolean? = null,
    @field:Size(min = 1) val defaultTargetDayTypes: Set<DayType>? = null,
)

data class CommuteRouteResponse(
    val id: Long,
    val name: String,
    val direction: CommuteDirection,
    val originLat: Double,
    val originLng: Double,
    val destinationLat: Double,
    val destinationLng: Double,
    val isActive: Boolean,
    /** KST 벽시계 (`HH:mm:ss`). 없으면 키째로 빠진다. */
    val defaultTargetArrivalTime: LocalTime?,
    /** enum 선언 순서(WEEKDAY, SATURDAY, SUNDAY_HOLIDAY)로 정렬. */
    val defaultTargetDayTypes: List<DayType>,
    val createdAt: Instant,
) {
    companion object {
        fun from(route: CommuteRoute) =
            CommuteRouteResponse(
                id = route.id!!,
                name = route.name,
                direction = route.direction,
                originLat = route.originLat,
                originLng = route.originLng,
                destinationLat = route.destinationLat,
                destinationLng = route.destinationLng,
                isActive = route.isActive,
                defaultTargetArrivalTime = route.defaultTargetArrivalTime,
                defaultTargetDayTypes = route.defaultTargetDayTypes.toSortedSet().toList(),
                createdAt = route.createdAt,
            )
    }
}

data class CommuteRouteDetailResponse(
    val route: CommuteRouteResponse,
    val legs: List<RouteLegResponse>,
)

data class RouteLegRequest(
    val id: Long? = null,
    @field:Positive val seqOrder: Int,
    val legType: LegType,
    @field:DecimalMin("-90.0") @field:DecimalMax("90.0") val startLat: Double? = null,
    @field:DecimalMin("-180.0") @field:DecimalMax("180.0") val startLng: Double? = null,
    @field:DecimalMin("-90.0") @field:DecimalMax("90.0") val endLat: Double? = null,
    @field:DecimalMin("-180.0") @field:DecimalMax("180.0") val endLng: Double? = null,
    @field:Positive val plannedDistanceM: Double? = null,
    val transitLineId: Long? = null,
    val boardStopId: Long? = null,
    val alightStopId: Long? = null,
    @field:Positive val plannedTravelSec: Int? = null,
)

data class ReplaceRouteLegsRequest(
    @field:NotEmpty @field:Valid val legs: List<RouteLegRequest>,
)

data class RouteLegResponse(
    val id: Long,
    val seqOrder: Int,
    val legType: LegType,
    val startLat: Double?,
    val startLng: Double?,
    val endLat: Double?,
    val endLng: Double?,
    val plannedDistanceM: Double?,
    val transitLineId: Long?,
    val boardStopId: Long?,
    val alightStopId: Long?,
    val plannedTravelSec: Int?,
    /**
     * TRANSIT 구간의 노선과 승하차 정류장 (#36). 앱이 "동탄 → 수서 (GTX-A)"를 띄우고 정류장 geofence를
     * 거는 데 쓴다. `…Id` 필드는 하위 호환으로 남긴다. WALK 구간은 null이라 응답에서 키째로 빠진다.
     */
    val transitLine: TransitLineResponse?,
    val boardStop: TransitStopResponse?,
    val alightStop: TransitStopResponse?,
    val signalCrossings: List<SignalCrossingResponse>,
) {
    companion object {
        fun from(
            leg: RouteLeg,
            crossings: List<RouteLegSignalCrossing>,
        ) = RouteLegResponse(
            id = leg.id!!,
            seqOrder = leg.seqOrder,
            legType = leg.legType,
            startLat = leg.startLat,
            startLng = leg.startLng,
            endLat = leg.endLat,
            endLng = leg.endLng,
            plannedDistanceM = leg.plannedDistanceM,
            transitLineId = leg.transitLine?.id,
            boardStopId = leg.boardStop?.id,
            alightStopId = leg.alightStop?.id,
            plannedTravelSec = leg.plannedTravelSec,
            transitLine = leg.transitLine?.let(TransitLineResponse::from),
            boardStop = leg.boardStop?.let { TransitStopResponse.from(it) },
            alightStop = leg.alightStop?.let { TransitStopResponse.from(it) },
            signalCrossings = crossings.map(SignalCrossingResponse::from),
        )
    }
}

data class SignalCrossingRequest(
    val trafficSignalId: Long,
    @field:Positive val seqOrder: Int,
    @field:NotBlank val approachDir: String,
    @field:NotBlank val signalKind: String = "Pd",
)

data class ReplaceSignalCrossingsRequest(
    @field:Valid val crossings: List<SignalCrossingRequest>,
)

data class SignalCrossingResponse(
    val id: Long,
    val trafficSignalId: Long,
    val seqOrder: Int,
    val approachDir: String,
    val signalKind: String,
) {
    companion object {
        fun from(crossing: RouteLegSignalCrossing) =
            SignalCrossingResponse(
                id = crossing.id!!,
                trafficSignalId = crossing.trafficSignal.id!!,
                seqOrder = crossing.seqOrder,
                approachDir = crossing.approachDir,
                signalKind = crossing.signalKind,
            )
    }
}