# desktop/openapi.json
{
  "components": {
    "schemas": {
...
      "CommuteRouteDetailResponse": {
        "properties": {
          "legs": {
            "items": {
              "$ref": "#/components/schemas/RouteLegResponse"
            },
            "type": "array"
          },
          "route": {
            "$ref": "#/components/schemas/CommuteRouteResponse"
          }
        },
        "required": [
          "legs",
          "route"
        ],
        "type": "object"
      },
      "CommuteRouteResponse": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "defaultTargetArrivalTime": {
            "type": [
              "string",
              "null"
            ]
          },
          "defaultTargetDayTypes": {
            "items": {
              "enum": [
                "WEEKDAY",
                "SATURDAY",
                "SUNDAY_HOLIDAY"
              ],
              "type": "string"
            },
            "type": "array"
          },
          "destinationLat": {
            "format": "double",
            "type": "number"
          },
          "destinationLng": {
            "format": "double",
            "type": "number"
          },
          "direction": {
            "enum": [
              "TO_WORK",
              "TO_HOME"
            ],
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "isActive": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "originLat": {
            "format": "double",
            "type": "number"
          },
          "originLng": {
            "format": "double",
            "type": "number"
          }
        },
        "required": [
          "createdAt",
          "defaultTargetDayTypes",
          "destinationLat",
          "destinationLng",
          "direction",
          "id",
          "isActive",
          "name",
          "originLat",
          "originLng"
        ],
        "type": "object"
      },
...
      "CreateCommuteRouteRequest": {
        "properties": {
          "defaultTargetArrivalTime": {
            "type": [
              "string",
              "null"
            ]
          },
          "defaultTargetDayTypes": {
            "items": {
              "enum": [
                "WEEKDAY",
                "SATURDAY",
                "SUNDAY_HOLIDAY"
              ],
              "type": "string"
            },
            "maxItems": 2147483647,
            "minItems": 1,
            "type": [
              "array",
              "null"
            ],
            "uniqueItems": true
          },
          "destinationLat": {
            "format": "double",
            "maximum": 90.0,
            "minimum": -90.0,
            "type": "number"
          },
          "destinationLng": {
            "format": "double",
            "maximum": 180.0,
            "minimum": -180.0,
            "type": "number"
          },
          "direction": {
            "enum": [
              "TO_WORK",
              "TO_HOME"
            ],
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "originLat": {
            "format": "double",
            "maximum": 90.0,
            "minimum": -90.0,
            "type": "number"
          },
          "originLng": {
            "format": "double",
            "maximum": 180.0,
            "minimum": -180.0,
            "type": "number"
          }
        },
        "required": [
          "destinationLat",
          "destinationLng",
          "direction",
          "isActive",
          "name",
          "originLat",
          "originLng"
        ],
        "type": "object"
      },
...
      "LegCalibrationResponse": {
        "properties": {
          "alightStopId": {
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "alightStopName": {
            "type": [
              "string",
              "null"
            ]
          },
          "boardStopId": {
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "boardStopName": {
            "type": [
              "string",
              "null"
            ]
          },
          "legType": {
            "enum": [
              "WALK",
              "TRANSIT"
            ],
            "type": "string"
          },
          "plannedDistanceM": {
            "format": "double",
            "type": [
              "number",
              "null"
            ]
          },
          "plannedTravelSec": {
            "format": "int32",
            "type": [
              "integer",
              "null"
            ]
          },
          "predictionRows": {
            "items": {
              "$ref": "#/components/schemas/PredictionCalibrationRowResponse"
            },
            "type": "array"
          },
          "routeLegId": {
            "format": "int64",
            "type": "integer"
          },
          "seqOrder": {
            "format": "int32",
            "type": "integer"
          },
          "transitLineId": {
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "transitLineName": {
            "type": [
              "string",
              "null"
            ]
          },
          "travelTimeRows": {
            "items": {
              "$ref": "#/components/schemas/TravelTimeCalibrationRowResponse"
            },
            "type": "array"
          },
          "walkingProfile": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/WalkingProfileResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "legType",
          "predictionRows",
          "routeLegId",
          "seqOrder",
          "travelTimeRows"
        ],
        "type": "object"
      },
...
      "PredictionCalibrationRowResponse": {
        "properties": {
          "biasSec": {
            "format": "int32",
            "type": "integer"
          },
          "dayType": {
            "enum": [
              "WEEKDAY",
              "SATURDAY",
              "SUNDAY_HOLIDAY"
            ],
            "type": "string"
          },
          "sampleCount": {
            "format": "int32",
            "type": "integer"
          },
          "stddevSec": {
            "format": "int32",
            "type": "integer"
          },
          "timeBandEnd": {
            "type": "string"
          },
          "timeBandStart": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "biasSec",
          "dayType",
          "sampleCount",
          "stddevSec",
          "timeBandEnd",
          "timeBandStart",
          "updatedAt"
        ],
        "type": "object"
      },
...
      "ReplaceRouteLegsRequest": {
        "properties": {
          "legs": {
            "items": {
              "$ref": "#/components/schemas/RouteLegRequest"
            },
            "minItems": 1,
            "type": "array"
          }
        },
        "required": [
          "legs"
        ],
        "type": "object"
      },
...
      "RouteCalibrationResponse": {
        "properties": {
          "globalWalkingProfile": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/WalkingProfileResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "legs": {
            "items": {
              "$ref": "#/components/schemas/LegCalibrationResponse"
            },
            "type": "array"
          },
          "minSamples": {
            "format": "int32",
            "type": "integer"
          },
          "routeId": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "legs",
          "minSamples",
          "routeId"
        ],
        "type": "object"
      },
      "RouteLegRequest": {
        "properties": {
          "alightStopId": {
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "boardStopId": {
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "endLat": {
            "format": "double",
            "maximum": 90.0,
            "minimum": -90.0,
            "type": [
              "number",
              "null"
            ]
          },
          "endLng": {
            "format": "double",
            "maximum": 180.0,
            "minimum": -180.0,
            "type": [
              "number",
              "null"
            ]
          },
          "id": {
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "legType": {
            "enum": [
              "WALK",
              "TRANSIT"
            ],
            "type": "string"
          },
          "plannedDistanceM": {
            "format": "double",
            "type": [
              "number",
              "null"
            ]
          },
          "plannedTravelSec": {
            "format": "int32",
            "type": [
              "integer",
              "null"
            ]
          },
          "seqOrder": {
            "format": "int32",
            "type": "integer"
          },
          "startLat": {
            "format": "double",
            "maximum": 90.0,
            "minimum": -90.0,
            "type": [
              "number",
              "null"
            ]
          },
          "startLng": {
            "format": "double",
            "maximum": 180.0,
            "minimum": -180.0,
            "type": [
              "number",
              "null"
            ]
          },
          "transitLineId": {
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          }
        },
        "required": [
          "legType",
          "seqOrder"
        ],
        "type": "object"
      },
      "RouteLegResponse": {
        "properties": {
          "alightStop": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/TransitStopResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "alightStopId": {
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "boardStop": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/TransitStopResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "boardStopId": {
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "endLat": {
            "format": "double",
            "type": [
              "number",
              "null"
            ]
          },
          "endLng": {
            "format": "double",
            "type": [
              "number",
              "null"
            ]
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "legType": {
            "enum": [
              "WALK",
              "TRANSIT"
            ],
            "type": "string"
          },
          "plannedDistanceM": {
            "format": "double",
            "type": [
              "number",
              "null"
            ]
          },
          "plannedTravelSec": {
            "format": "int32",
            "type": [
              "integer",
              "null"
            ]
          },
          "seqOrder": {
            "format": "int32",
            "type": "integer"
          },
          "signalCrossings": {
            "items": {
              "$ref": "#/components/schemas/SignalCrossingResponse"
            },
            "type": "array"
          },
          "startLat": {
            "format": "double",
            "type": [
              "number",
              "null"
            ]
          },
          "startLng": {
            "format": "double",
            "type": [
              "number",
              "null"
            ]
          },
          "transitLine": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/TransitLineResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "transitLineId": {
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          }
        },
        "required": [
          "id",
          "legType",
          "seqOrder",
          "signalCrossings"
        ],
        "type": "object"
      },
...
      "SignalCrossingResponse": {
        "properties": {
          "approachDir": {
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "seqOrder": {
            "format": "int32",
            "type": "integer"
          },
          "signalKind": {
            "type": "string"
          },
          "trafficSignalId": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "approachDir",
          "id",
          "seqOrder",
          "signalKind",
          "trafficSignalId"
        ],
        "type": "object"
      },
...
      "TransitLineResponse": {
        "properties": {
          "agency": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "externalId": {
            "type": [
              "string",
              "null"
            ]
          },
          "hasRealtimeApi": {
            "type": "boolean"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "mode": {
            "enum": [
              "BUS",
              "SUBWAY",
              "GTX"
            ],
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "stdgCd": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "createdAt",
          "hasRealtimeApi",
          "id",
          "mode",
          "name"
        ],
        "type": "object"
      },
      "TransitStopResponse": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "distanceM": {
            "format": "double",
            "type": [
              "number",
              "null"
            ]
          },
          "externalId": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "lat": {
            "format": "double",
            "type": "number"
          },
          "lng": {
            "format": "double",
            "type": "number"
          },
          "mode": {
            "enum": [
              "BUS",
              "SUBWAY",
              "GTX"
            ],
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "stdgCd": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "createdAt",
          "id",
          "lat",
          "lng",
          "mode",
          "name"
        ],
        "type": "object"
      },
      "TravelTimeCalibrationRowResponse": {
        "properties": {
          "dayType": {
            "enum": [
              "WEEKDAY",
              "SATURDAY",
              "SUNDAY_HOLIDAY"
            ],
            "type": "string"
          },
          "meanSec": {
            "format": "int32",
            "type": "integer"
          },
          "sampleCount": {
            "format": "int32",
            "type": "integer"
          },
          "stddevSec": {
            "format": "int32",
            "type": "integer"
          },
          "timeBandEnd": {
            "type": "string"
          },
          "timeBandStart": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "dayType",
          "meanSec",
          "sampleCount",
          "stddevSec",
          "timeBandEnd",
          "timeBandStart",
          "updatedAt"
        ],
        "type": "object"
      },
...
      "UpdateCommuteRouteRequest": {
        "properties": {
          "clearDefaultTargetArrivalTime": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "defaultTargetArrivalTime": {
            "type": [
              "string",
              "null"
            ]
          },
          "defaultTargetDayTypes": {
            "items": {
              "enum": [
                "WEEKDAY",
                "SATURDAY",
                "SUNDAY_HOLIDAY"
              ],
              "type": "string"
            },
            "maxItems": 2147483647,
            "minItems": 1,
            "type": [
              "array",
              "null"
            ],
            "uniqueItems": true
          },
          "isActive": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "name": {
            "pattern": "(?s).*\\S.*",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "type": "object"
      },
...
      "WalkingProfileResponse": {
        "properties": {
          "avgSpeedMps": {
            "format": "double",
            "type": "number"
          },
          "sampleCount": {
            "format": "int32",
            "type": "integer"
          },
          "stddevSpeedMps": {
            "format": "double",
            "type": "number"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "avgSpeedMps",
          "sampleCount",
          "stddevSpeedMps",
          "updatedAt"
        ],
        "type": "object"
      }
    }
  },
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/commute-routes": {
      "get": {
        "operationId": "list",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CommuteRouteResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "commute-route-controller"
        ]
      },
      "post": {
        "operationId": "create_4",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommuteRouteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommuteRouteResponse"
                }
              }
            },
            "description": "Created"
          }
        },
        "tags": [
          "commute-route-controller"
        ]
      }
    },
    "/api/v1/commute-routes/{id}": {
      "get": {
        "operationId": "detail",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommuteRouteDetailResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "commute-route-controller"
        ]
      },
      "patch": {
        "operationId": "update_1",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCommuteRouteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommuteRouteResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "commute-route-controller"
        ]
      }
    },
    "/api/v1/commute-routes/{id}/calibration": {
      "get": {
        "operationId": "calibration",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RouteCalibrationResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "commute-route-controller"
        ]
      }
    },
    "/api/v1/commute-routes/{id}/legs": {
      "put": {
        "operationId": "replaceLegs",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceRouteLegsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommuteRouteDetailResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "commute-route-controller"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://localhost"
    }
  ]
}